Add features - Add auto close search panel component when clicked out component area - Change header text color
