+++ src/components/core/LinkIconButton.jsx
@@ -6,7 +6,7 @@ import { Link } from "react-router-dom";
 
 const LinkIconButton = ({ text, icon, to }) => {
     const [btnBG, setBtnBG] = useState('');
-    const [textColor, setTextColor] = useState('text-gray-700');
+    const [textColor, setTextColor] = useState('text-green-900');
 
     const handleFocusGained = () => {
         setBtnBG('bg-green-900');
@@ -20,7 +20,7 @@ const LinkIconButton = ({ text, icon, to }) => {
     return (
       <Link
         to={to}
-        className={`flex items-center justify-center font-bold text-xsm md:text-sm transition-all pt-1 pb-1 pl-1 pr-3 rounded-3xl ${btnBG}`}
+        className={`flex items-center justify-center gap-1 font-bold text-xsm md:text-sm transition-all pt-1 pb-1 pl-1 pr-3 rounded-3xl ${btnBG}`}
         onMouseEnter={handleFocusGained}
         onMouseLeave={handleFocusLost}
       >
